moviedb/requirements.txt
requests
pytest
vcrpy